requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning & NLP
torch==2.6.0
transformers==4.35.0
google-generativeai==0.3.0

# Data Validation & Processing
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6

# Additional utilities (if needed)
numpy>=1.21.0
regex>=2023.0.0